requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.6
django-cors-headers==3.1.1
djangorestframework==3.10.3
gunicorn==20.0.4
idna==2.8
mysqlclient==1.4.4
Pillow==6.2.1
pytz==2019.3
requests==2.22.0
sqlparse==0.3.0
urllib3==1.25.7
